Add files via upload
A JavaScript on our group project from the interaction design course Tangible embodied interactions. It uses a Tensorflow library to control the tracking of the camera and the artefacts. It tracks the pictures of a banana, clock, and stop sign that have been printed on cubes, to be able to control sound parameters such as pitch and speed depending on the movement.